Clean up MakeInstaller script after debug

+++ tools/Scripts/MakeInstaller.py
@@ -233,16 +233,10 @@ def osDependentPreparation():
         Functions.printNeutralMessage(f'No preparation needed for os {CONFIG.os}')
 
 def installQtInstallerFramework():
-    #if os.path.exists(qtifwDirPath()):
-    #    Functions.printNeutralMessage(f'QtInstallerFramework was already installed to {qtifwDirPath()}')
-    #    return
+    if os.path.exists(qtifwDirPath()):
+        Functions.printNeutralMessage(f'QtInstallerFramework was already installed to {qtifwDirPath()}')
+        return
     try:
-        print("========================== 1")
-        print("----------------------- 0")
-        Functions.run('ls', '~')
-        print("----------------------- 1")
-        Functions.run('ls', '/home/runner')
-
         message = f'install QtInstallerFramework to {qtifwDirPath()}'
         silent_script = os.path.join(CONFIG.scripts_dir, CONFIG['ci']['scripts']['silent_install'])
         Functions.installSilently(
@@ -250,21 +244,6 @@ def installQtInstallerFramework():
             silent_script=silent_script
         )
         time.sleep(30)
-
-        print("========================== 2")
-        print("----------------------- 0")
-        Functions.run('ls', '~')
-        print("----------------------- 1")
-        Functions.run('ls', '/home/runner')
-        print("----------------------- 2")
-        Functions.run('ls', '/home/runner/Qt')
-        print("----------------------- 3")
-        Functions.run('ls', '/home/runner/Qt/QtIFW-4.4.1')
-        print("----------------------- 4")
-        Functions.run('ls', '/home/runner/Qt/QtIFW-4.4.1/bin')
-        print("----------------------- 5")
-
-
     except Exception as exception:
         Functions.printFailMessage(message, exception)
         sys.exit(1)
@@ -328,7 +307,7 @@ def createInstallerSourceDir():
 
 def createOfflineInstaller():
     try:
-        message = 'create installer'
+        message = 'create offline installer'
         qtifw_bin_dir_path = os.path.join(qtifwDirPath(), 'bin')
         qtifw_binarycreator_path = os.path.join(qtifw_bin_dir_path, 'binarycreator')
         qtifw_installerbase_path = os.path.join(qtifw_bin_dir_path, 'installerbase')
